+++ backend/aiconsole/core/assets/materials/db_crud.py
@@ -1,70 +1,13 @@
 from fastapi import HTTPException
-from pydantic import BaseModel, Field, validator
-from typing import List, Optional
 
-from sqlalchemy.orm import sessionmaker, Session
+from sqlalchemy.orm import Session
 from aiconsole.core.assets.types import AssetLocation
-from aiconsole.core.assets.materials.db_model import DbMaterial
+from aiconsole.core.assets.materials.db_model import DbMaterial, DbMaterialSchema, DbMaterialUpdateSchema
 
-# Pydantic schemas were required by the environment 
-# Also Pydantic Models make endpoints developement more managable and scalable
-# ChatGPT advised me to create separate schema for patch method. 
-# All fields are optional there and only non-None are used to update
-
-class DbMaterialSchema(BaseModel):
-    id: Optional[str] = Field(None, description="Primary key")
-    name: str = Field(..., description="Unique name of the material")
-    version: str
-    usage: Optional[str] = None
-    usage_examples: Optional[str] = None
-    defined_in: Optional[str] = None
-    type: Optional[str] = 'material'
-    default_status: str
-    status: str
-    content_type: str = Field(..., description="Type of the material")
-    content: str
-
-    class Config:
-        from_attributes = True
-
-    @validator('content_type')
-    def validate_content_type(cls, value):
-        allowed_types = ['static_text', 'dynamic_text', 'api']
-        if value not in allowed_types:
-            raise ValueError(f"Invalid content_type: '{value}'. Must be one of {allowed_types}.")
-        return value
-
-    @validator('id', always=True)
-    def validate_id(cls, value):
-        if value is None:
-            raise ValueError("id must be provided.")
-        return value
-
-    @validator('name')
-    def validate_name(cls, value):
-        if len(value) == 0:
-            raise ValueError("name must not be empty.")
-        return value
-    
-class DbMaterialUpdateSchema(BaseModel):
-    id: Optional[str] = None
-    name: Optional[str] = None
-    version: Optional[str] = None
-    usage: Optional[str] = None
-    usage_examples: Optional[str] = None
-    defined_in: Optional[str] = None
-    type: Optional[str] = None
-    default_status: Optional[str] = None
-    status: Optional[str] = None
-    content_type: Optional[str] = None
-    content: Optional[str] = None
-
-    class Config:
-        from_attributes = True
 
 # CRUD operations
 def _create_material_db(db: Session, material: DbMaterialSchema):
-    db_material = DbMaterial(**material.model_dump())
+    db_material = DbMaterial(**material.model_dump_filtered(DbMaterial))
     db.add(db_material)
     db.commit()
     db.refresh(db_material)
@@ -134,3 +77,24 @@ def _material_exists(db: Session, location:AssetLocation, material_id: str) -> b
         material = _get_material_db(db, material_id)
 
     return (material is not None)
+
+# def _preview_material(material: Material):
+#     content_context = ContentEvaluationContext(
+#         chat=Chat(
+#             id="chat",
+#             name="",
+#             last_modified=datetime.now(),
+#             title_edited=False,
+#             message_groups=[],
+#         ),
+#         agent=create_user_agent(),
+#         gpt_mode=SPEED_GPT_MODE,
+#         relevant_materials=[],
+#     )
+
+#     try:
+#         rendered_material = await material.render(content_context)
+#     except ValueError as e:
+#         return JSONResponse(e.args[1].model_dump(exclude_none=True))
+
+#     return JSONResponse(rendered_material.model_dump(exclude_none=True))
